Edit task and add task panels finishe, some presenter and controller to be implemented later
