+++ src/main/java/gui/new_task/NewTaskController.java
@@ -6,7 +6,7 @@ public class NewTaskController extends TaskController {
     public NewTaskController(Boolean log_messages) {
         super(log_messages);
     }
-    public void add_habit(int habit_id, String name, String description, String frequency){
-        print_log_message("adding habit {ID: " + habit_id + " name: [" + name + "] description: [" + description + "] frequency: [" + frequency + "]}");
+    public void add_habit(String name, String description, String frequency){
+        print_log_message("adding habit {ID: [unassigned] name: [" + name + "] description: [" + description + "] frequency: [" + frequency + "]}");
     }
 }
