+++ src/main/java/gui/new_task/NewTaskView.java
@@ -1,0 +1,33 @@
+package gui.new_task;
+import gui.task.TaskView;
+
+import javax.swing.*;
+import java.awt.*;
+import java.awt.event.ActionEvent;
+import java.awt.event.ActionListener;
+
+public class NewTaskView extends TaskView {
+    private NewTaskController newTaskController;
+    private JButton add;
+
+    public NewTaskView(NewTaskController newTaskController) {
+        super("New Task",   newTaskController);
+        this.newTaskController = newTaskController;
+        createUI(800, 600, view_name);
+        addAddButton();
+        addCancelButton();
+    }
+
+    private void addAddButton(){
+        add = getSizedButton(100, 36, "Save", new Color(67, 182, 240));
+        add.addActionListener(new ActionListener() {
+            public void actionPerformed(ActionEvent e) {
+                String habit_name = habitNameTF.getText();
+                String habit_desc = habitDescTF.getText();
+                String habit_repeat = habitRepeatCB.getSelectedItem().toString();
+                newTaskController.add_habit(habit_name, habit_desc, habit_repeat);
+            }
+        });
+        buttonsRow.add(add);
+    }
+}
